+++ admin/updates/20190619-similarity.sql
@@ -37,24 +37,4 @@ ALTER TABLE similarity_stats
   FOREIGN KEY (metric)
   REFERENCES similarity_metrics (metric);
 
-
-CREATE TABLE similarity_eval (
-  user_id INTEGER, -- FK to user
-  query_mbid UUID,
-  result_mbids UUID[],
-  metric TEXT, -- FK to metric
-  rating SMALLINT,
-  suggestion TEXT
-);
-
-ALTER TABLE similarity_eval
-  ADD CONSTRAINT similarity_eval_fk_user
-  FOREIGN KEY (user_id)
-  REFERENCES "user" (id);
-
-ALTER TABLE similarity_eval
-  ADD CONSTRAINT similarity_eval_fk_metric
-  FOREIGN KEY (metric)
-  REFERENCES similarity_metrics (metric);
-
 COMMIT;
